Extracted math funcs.
